Convert gen_loader_hikey.py and gen_loader_hikey960.py to Python 3
Python 2 is deprecated. On many distributions the python command is a
symbolic link to python3 and on some of them it isn't even possible to
make it a link to python2 using a supported method (Ubuntu used to
provide package python-is-python2 but it doesn't anymore). In a not too
distant future, some distributions will remove Python 2 altogether.

Therefore:
- Update the gen_loader_hikey.py and gen_loader_hikey960.py to Python 3.
This was done with the help of the 2to3 command and a few manual fixes.
- Invoke Python as python3 not python.

From now on, users are expected to have a working python3 command to use
this project.

Tested in the OP-TEE OS development environment for HiKey 620 [1] and
HiKey 960 [2] with 'make', 'make flash' and 'make recovery'.

Link: [1] https://optee.readthedocs.io/en/latest/building/devices/hikey620.html#build-instructions
Link: [2] https://optee.readthedocs.io/en/latest/building/devices/hikey960.html#build-instructions
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>